Toast notif working
