+++ frontend/js/components/toast/toast.html
@@ -1,8 +1,8 @@
-<div class="toast" role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="30">
+<div class="toast" role="alert" aria-live="assertive" aria-atomic="true" data-bs-delay="30000">
     <div class="toast-header">
     <img width="50" height="50" src="https://png.pngtree.com/png-clipart/20190904/ourmid/pngtree-80-3d-text-png-image_18456.jpg" class="rounded me-2" alt="...">
     <strong class="me-auto">Bootstrap</strong>
-    <small id="timeSinceToast" class="text-muted">just now</small>
+    <small id="timeSinceToast" class="text-muted"></small>
     <button type="button" class="btn-close" data-bs-dismiss="toast" aria-label="Close"></button>
     </div>
     <div class="toast-body">
